+++ packages/_config/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @bepp/config
 
+## 1.2.9
+
+### Patch Changes
+
+- update docker container
+
 ## 1.2.8
 
 ### Patch Changes