+++ packages/app/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @bepp/app
 
+## 1.3.11
+
+### Patch Changes
+
+- update docker container
+
 ## 1.3.10
 
 ### Patch Changes