+++ packages/bepp/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @bepp/bepp
 
+## 1.3.11
+
+### Patch Changes
+
+- update docker container
+
+- Updated dependencies []:
+  - @bepp/core@1.3.11
+
 ## 1.3.10
 
 ### Patch Changes